requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
requests>=2.32.0
tqdm>=4.66.0
python-dotenv>=1.0.1

# NLP / RAG
jieba>=0.42.1
rank-bm25>=0.2.2
sentence-transformers>=3.1.0
transformers>=4.44.0
torch>=2.2.0             # CPU版；如需GPU请自行安装对应CUDA版
faiss-cpu>=1.8.0.post1   # Windows/Mac/Linux 通用轮子；若装不上则临时注释

# OpenAI / 兼容客户端
openai>=1.44.0

# Optional（数值）
numpy>=1.26.0

peft

hf_xet